Unify HPackIntegerTest
